listen on a particular interface
By default restricts to only listen on localhost so the
user will need to use an ssh proxy to the instance. This
provides better security and auditing. If needed the host
system property can be set to allow connecting directly
to jmx after opening up the port in the instance firewall.