name the Configuration binding
There are a number of apps internally that create a binding
of `Configuration` to something other than the main config.
Typically this is a binding for explorers with a small set
of properties. This change specifies a name for the binding
we care about to avoid the conflict.

For our use-cases this will always be an optional injection
and will get the static archaius instance as a fallback.