allow for config specific to environment
For running outside of the main account there are a handful
of properties where we need different settings. The need for
this should get reduced over time as we make things more
consistent.

Also sets up aliases for common env settings with defaults
so it is easier to run locally.