make it easier to send emails with attachments
Adds a helper lib abstracted out of internal alert backend
to make it easier to construct the `RawMessage` type needed
by SES to send an email with attachments.

The common case is a basic HTML email message with some
images such as Atlas graphs providing alert context. Also
used for canary reports and a handful of other places.