remove rxnetty metrics module (#270)
We typically use the http request level metrics to
understand what the clients are doing. The lower level
rxnetty metrics aren't used much and don't follow
the recommended conventions so we would prefer not
to publish them for our apps by default.

An app owner can explicitly enable if they are still
desirable in certain use-cases.