switch to using builder for AWS clients
Fixes some deprecation warnings with recent AWS SDK by
switching to use the builders for clients. This change
also allows us to simplify the config because we only
need to know the region, not the full endpointn that we
need to construct. The config has been simplified to
only include overrides for services like SES and Route53
where the service is not available in all of the regions
where we need to access them.