add module with basic support for aws sdk 2.x
Provides a module that works with the 2.x SDK. This is
experimental until the 2.x SDK is released and out of
developer preview mode. For now it is purely for early
adopters at Netflix to try out and provide feedback.

Note, jdk9 build is disabled until there is a fix of
either jackson 2.9.1 or in jdk9 itself for manifest
files that contain backticks. See:

https://bugs.openjdk.java.net/browse/JDK-8186334