change packages for iep-module-admin
When using the new java module system, classes for a
given package should only be in one jar. This change
updates the iep-module-admin package so it does not
overlap with iep-admin. The guice module package was
changed as it is typically auto-loaded via the
ServiceLoader, so even though this is a breaking change
it is not expected to break any code in practice.