add ConfigManager utility with iep loading logic (#353)
This makes it easier to reuse the loading we are using
for the typesafe config objects and in most cases can
be used as a substitute for `ConfigFactory.load` to make
those use-cases consistent. Unfortunately there are some
places that do not get the injected config and this helps
avoid inconsistencies.