use global registry internal to config
If the registry implementation needs to use the config, then
it will create a circular dependency between the config and
the registry instance and fail with:

```
Caused by: java.lang.IllegalStateException: This is a proxy
  used to support circular references. The object we're
  proxying is not constructed yet. Please wait until after
  injection has completed to use this object.
```

To avoid this the config bindings will now use the Spectator
global registry internally. This breaks the cycle and still
allows the remote config loading to report metrics as long
as the registry binding adds itself to the global registry.