improve support for using sending authorization (#367)
When using SES sending authorization and SendRawEmail, the
SourceArn will only work if the From address header is not
specified in the raw message. This is not clear from reading
the AWS documentation.

This change updates the helper to allow the FromArn to be
specified as part of the raw message which does work when
a from header is included. It also allows the from header
to be omitted now so that it can be included by calling
withSource on the SendRawEmailRequest object.

The readme has also been updated to cover the use with sending
authorization.