update iep-ses to work with aws-sdk-java-v2
Adds helpers to get the `ByteBuffer` directly so it can
be used to construct the `RawMessage` object for either
v1 or v2 of the SDK. It is now fairly straightforward
to use with either version, but there is an explicit
dependency on v1 so it will get pulled in even if the
user is using v2.

Methods that refer to v1 classes have been deprecated so
we can remove the explicit dependency in iep-ses 2.0. At
that point the only version of the AWS libraries in the
classpath will be what the user depends on.