fix endless loop for listHostedZones
This result object has both `withMarker` and `withNextMarker`.
The marker is what was used for the request that generated the
result and should get ignored for pagination. Before, if the
call resulted in multiple pages, the marker would always have
a value for the last page and cause it to indicate that there
was another request. However it was correctly updating the
request with the result of next marker so it would start over
from the beginning.

The test case has been updated to reproduce the issue.