userservice: fix cache age metric (#401)
A reference to the AtomicLong for the PolledMeter was not
retained so it would get garbage collected and stop reporting.

Also removes an unnecessary print stack trace call and updates
the client to set the endpoint for metrics. Since the remote
service in this case does not have the headers set, it is not
picked up automatically.