add userservice based on HTTP GET check (#433)
Adds a new implementation of UserService that does an
inline call to check if an email is valid. This works
for the newer API internally and avoids the need to
cache the full set of users.

To reduce the cost of repeated checks it will cache
the result of checks for a period of time. Set to a
smaller value to be more responsive to new addresses.

The behavior can also be customized for a failure to
talk to the remote service. If it is a nice to have
check, then it might be better to treat it as valid
until it can be confirmed. Otherwise, an outage of
the remote service will prevent validation unless the
result has already been cached.