support dynamic override for ConfigManager
Adds support for setting a dynamic override layer for the
ConfigManager. This can be used for updating the config
based on runtime sources such as the command line or a
remote property service.

This should allow us to phase out the usage of the
`atlas-config` library and simplify the overall config story.