deprecate archaius2 module (#460)
Deprecates the archaius2 module to make the config story
more clear and just standardize on Typesafe Config.

For cases where dynamic configuration is desirable a new
module `iep-module-dynconfig` has been added that works
with the ConfigManager from `iep-nflxenv`. It uses the
same properties and formats as the old modules so it can
easily be dropped in as long as you do not rely on the
Archaius config interfaces. This module also has an admin
endpoint for `/props` so it is easy to inspect the config
from the admin interface just like before.