lazily access typesafe config
Updates NetflixEnvironment to only access the ConfigManager,
and thus load the typesafe config, if one of the helper
methods that rely on it are called. This avoids static
initialization errors for the NetflixEnvironment class if
there is a bad `application.conf` or `reference.conf` file
in the classpath. This has two benefits:

- For apps that are just relying on telemetry clients that
  use the library for getting common tags, the other helpers
  will not be accessed and the issues is avoided entirely.
- For apps that do access the methods, the exception will be
  thrown to the caller and should be easier to diagnose than
  `java.lang.NoClassDefFoundError: Could not initialize class
  com.netflix.iep.NetflixEnvironment`.