try to show status for nimble groups
Nimble groups report to Eureka as if they are already
part of the transplant server group. Before they were
mostly ignored and would always have a status of not
registered. This change adds some hacks to try and
find the actual status for the nimble registration and
use that for the merged group. This only works when
there is a source such as Edda that has the correct
information earlier on.