add mantis tags to default atlas tag keys
The Mantis tags have long been a part of Atlas metrics common tagging, and
they only appear in the Mantis application clusters, so it is safe to add
them to the default set.

Making this change here, rather than creating a separate set of keys to
cover this use case, means that there will remain one standard source for
Atlas metrics common tags.