fix memory leak for compressed responses (#518)
The server groups loader could leak native memory if
the gzip stream wasn't closed. Since the json library
will automatically try to close if it hits the end
of stream, the leak would only occur if there was an
error parsing or the parsing stopped before fully
consuming the input.